+++ ui/src/app/app-routing.module.ts
@@ -1,14 +1,14 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { ComicsComponent } from './comics/comics.component';
+import { LibraryComponent } from './library/library.component';
 import { ReaderComponent } from './reader/reader.component';
 import { PreferencesComponent } from './preferences/preferences.component';
 import { FullScreenReaderComponent } from './full-screen-reader/full-screen-reader.component';
 
 const routes: Routes = [
-  { path: '', redirectTo: '/comics', pathMatch: 'full' },
-  { path: 'comics', component: ComicsComponent },
+  { path: '', redirectTo: '/library', pathMatch: 'full' },
+  { path: 'library', component: LibraryComponent },
   { path: 'read/:id/:page', component: ReaderComponent },
   { path: 'read-full-screen/:id/:page', component: FullScreenReaderComponent },
   { path: 'preferences', component: PreferencesComponent }
