+++ ui/src/app/library/library.component.spec.ts
@@ -1,18 +1,18 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { TestModule } from './../test.module';
 
-import { ComicsComponent } from './comics.component';
+import { LibraryComponent } from './library.component';
 
-describe('ComicsComponent', () => {
-  let component: ComicsComponent;
-  let fixture: ComponentFixture<ComicsComponent>;
+describe('LibraryComponent', () => {
+  let component: LibraryComponent;
+  let fixture: ComponentFixture<LibraryComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule(TestModule).compileComponents();
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(ComicsComponent);
+    fixture = TestBed.createComponent(LibraryComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
