+++ ui/src/app/library/scanner/scanner.component.html
