+++ ui/src/app/library/scanner/scanner.component.spec.ts
